src/server/requirements.txt
Django==2.0.7
django-extensions==2.1.0
django-fsm==2.6.0
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
psycopg2-binary==2.7.5
pytz==2018.5
flake8==3.5.0
django-allauth==0.36.0
jsondiff==1.1.2
django-debug-toolbar==1.9.1
schema==0.6.8
